Nightly Build June 3, 2015
Fixed Container background rendering (background and object are
rendered separately).
Implemented how buttons will work.
Game class contains examples of all current GUI elements which will
appear when the game is launched.
Added ColorCodes class for pretty custom error outputting.

+++ core/src/com/icnhdevelopment/wotn/handlers/CInputProcessor.java
@@ -46,7 +46,7 @@ public class CInputProcessor implements InputProcessor {
         return (!buttonRight && lastButtonRight);
     }
 
-    public boolean mouseHovered(int x, int y, int width, int height) {
+    public boolean mouseHovered(float x, float y, float width, float height) {
         return (new Rectangle(x, y, width, height).contains(new Rectangle(mousePosition.x, mousePosition.y, 1, 1)));
     }
